Move util functions to utils
